pyproject.toml
[project]
name = "veo-soccer-analysis"
version = "0.1.0"
description = "Local-first soccer match video analysis for Apple Silicon"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pydantic>=2.5.0",
    "pyyaml>=6.0",
    "opencv-python>=4.8.0",
    "ffmpeg-python>=0.2.0",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "ultralytics>=8.0.0",
    "filterpy>=1.4.5",
    "scikit-learn>=1.3.0",
    "scipy>=1.11.0",
    "pyarrow>=14.0.0",
    "fastparquet>=2023.10.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "jinja2>=3.1.2",
    "click>=8.1.0",
    "rich>=13.0.0",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "ipython>=8.17.0",
]

[project.scripts]
veo-analyze = "src.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_backend"
